watchdog/requirements.txt
backports.ssl==0.0.9
beautifulsoup4==4.6.0
cffi==1.5.0
click==6.7
cryptography==1.2.2
cssselect==1.0.3
enum34==1.1.2
funcsigs==0.4
hupper==1.1
idna==2.0
IMAPClient==2.0.0
ipaddress==1.0.16
lxml==4.2.1
mechanize==0.2.5
mock==1.3.0
PasteDeploy==1.5.2
pbr==1.8.1
plaster==1.0
plaster-pastedeploy==0.5
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==0.15.1
pyquery==1.4.0
pyramid==1.9.1
pyramid-mailer==0.15.1
pytz==2018.4
py-zabbix==1.1.3
repoze.lru==0.7
repoze.sendmail==4.4.1
six==1.11.0
transaction==2.2.1
translationstring==1.3
venusian==1.1.0
virtualenv==15.2.0
waitress==1.1.0
WebOb==1.8.1
WebTest==2.0.29
WSGIProxy2==0.4.4
zope.cachedescriptors==4.3.1
zope.deprecation==4.3.0
zope.event==4.3.0
zope.interface==4.5.0
zope.schema==4.5.0
zope.testbrowser==5.2.4
